Fix #37: Cannot count(*) on an index
The issue was specifically for indexes that don't project all attributes. The code was treating a "select expression with 'None' fields" as selecting *all* fields (inferring "select *"). Unfortunately doing "select count(*)" also will produce 'None' fields. There's a lot wrong with these assumptions and the structure of the data, but for now I just added a special case for when the expression is just selecting the count.